+++ XamTwitch/XamTwitch/Services/ITwitchAnonymousHttpService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+using XamTwitch.Models;
+
+namespace XamTwitch.Services
+{
+    public interface ITwitchAnonymousHttpService
+    {
+        Task<TwitchToken> GetTwitchTokenAsync(string userName);
+    }
+}
