+++ XamTwitch/XamTwitch/Services/ITwitchHttpService.cs
@@ -8,6 +8,5 @@ namespace XamTwitch.Services
     {
         Task<TwitchGames> GetTwitchGamesAsync(string gameName);
         Task<TwitchStreams> GetTwitchStreamsAsync();
-        Task<TwitchToken> GetTwitchTokenAsync(string userName);
     }
 }
