+++ XamTwitch/XamTwitch/ViewModels/PlayerPageViewModel.cs
@@ -1,12 +1,7 @@
 ﻿using System;
-using System.Net.Http;
-using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.MobCAT;
 using Microsoft.MobCAT.MVVM;
-using Newtonsoft.Json;
-using Xamarin.Essentials;
 using XamTwitch.Helpers;
 using XamTwitch.Models;
 using XamTwitch.Services;
@@ -15,7 +10,7 @@ namespace XamTwitch.ViewModels
 {
     public class PlayerPageViewModel : BaseNavigationViewModel
     {
-        private readonly ITwitchHttpService _twitchHttpService;
+        private readonly ITwitchAnonymousHttpService _twitchAnonymousHttpService;
         private readonly ITwitchPlaylistHttpService _twitchPlaylistHttpService;
 
         private TwitchStream _stream;
@@ -40,7 +35,7 @@ namespace XamTwitch.ViewModels
 
         public PlayerPageViewModel()
         {
-            _twitchHttpService = ServiceContainer.Resolve<ITwitchHttpService>();
+            _twitchAnonymousHttpService = ServiceContainer.Resolve<ITwitchAnonymousHttpService>();
             _twitchPlaylistHttpService = ServiceContainer.Resolve<ITwitchPlaylistHttpService>();
         }
 
@@ -49,7 +44,7 @@ namespace XamTwitch.ViewModels
             if (Stream == null)
                 return;
 
-            var token = await _twitchHttpService.GetTwitchTokenAsync(Stream.UserName);
+            var token = await _twitchAnonymousHttpService.GetTwitchTokenAsync(Stream.UserName);
             var streamUrl = await _twitchPlaylistHttpService.GetPlaylistUriAsync(Stream.UserName, token);
             StreamSource = streamUrl;
         }
